wip: retrieve function secrets via Kafka RPC in RequestReaderService
