+++ src/MangoFaaS.Models/FunctionSecretsRpc.cs
@@ -1,0 +1,17 @@
+namespace MangoFaaS.Models;
+
+public class FunctionSecretsRequest
+{
+    public Guid FunctionId { get; set; }
+}
+
+public class FunctionSecretsResponse
+{
+    public List<FunctionSecretEntry> Secrets { get; set; } = [];
+}
+
+public class FunctionSecretEntry
+{
+    public required string Name { get; set; }
+    public required string Value { get; set; }
+}
