+++ src/MangoFaaS.Secrets/Dto/FunctionSecretsRpc.cs
@@ -1,17 +1,2 @@
-namespace MangoFaaS.Secrets.Dto;
-
-public class FunctionSecretsRequest
-{
-    public Guid FunctionId { get; set; }
-}
-
-public class FunctionSecretsResponse
-{
-    public List<FunctionSecretEntry> Secrets { get; set; } = [];
-}
-
-public class FunctionSecretEntry
-{
-    public required string Name { get; set; }
-    public required string Value { get; set; }
-}
+// RPC DTOs (FunctionSecretsRequest, FunctionSecretsResponse, FunctionSecretEntry)
+// have been moved to MangoFaaS.Models for cross-service sharing.
